pyproject.toml
[tool.poetry]
name = "atsume"
version = "0.2.2"
description = ""
authors = ["Peter DeVita <mewtwo2643@yahoo.com>"]
readme = "Readme.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
hikari = "^2.0.0.dev120"
click = "^8.1, <8.1.4"
hikari-tanjun = "^2.15.0"
aiohttp = "^3.8.4"
ormar = "^0.12.2"
alembic = "^1.11.1"
asyncpg = { version = "^0.28.0", optional = true }
psycopg2-binary = { version = "^2.9.6", optional = true }
aiopg = { version = "^1.4.0", optional = true }
aiomysql = { version = "^0.2.0", optional = true }
pymysql = { version = "^1.1.0", optional = true }
aiosqlite = { version = "^0.19.0", optional = true }
hupper = "^1.12"

[tool.poetry.group.dev.dependencies]
myst-parser = "^2.0.0"
sphinx-autodoc2 = "^0.4.2"
sphinx-autoapi = "^2.1.1"
sphinxcontrib-asyncio = "^0.3.0"
mypy = "^1.4.1"
black = "^23.7.0"
pre-commit = "^3.3.3"
nox = "^2023.4.22"
nox-poetry = "^1.0.3"
sqlalchemy-stubs = "^0.4"
sphinx = "^7.0.1"
sphinx-reload = "^0.2.0"
furo = "^2023.5.20"
towncrier = "^23.6.0"

[tool.poetry.extras]
postgresql = ["asyncpg", "psycopg2-binary"]
postgres = ["asyncpg", "psycopg2-binary"]
aiopg = ["aiopg", "psycopg2-binary"]
mysql = ["aiomysql", "PyMySQL"]
sqlite = ["aiosqlite"]

[tool.poetry.scripts]
atsume = "atsume.cli.project:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.towncrier]
package = "hikari-atsume"
package_dir = "atsume"
filename = "CHANGELOG.md"
